refactor(cli) : Gallery object now uses GalleryToml class
This refactor still leaves large parts of the application basically using
gallery_toml dictionary and passing it around. Hopefully these will disappear
in coming refactoring rounds.

Also, a better name for GalleryToml object is needed. I feel toml file is just
an implementation detail, however important but still. The actual point here is
that this class is a representation of flat file clear/plaintext database of all
the images